requirements.txt
# API Gateway
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.2
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Agents & Task Queue
celery==5.3.6
redis==5.0.1
openai==1.3.5
anthropic==0.7.7
requests==2.31.0
textblob==0.17.1
transformers==4.35.2
torch>=2.6.0

# Discord Bot
discord.py==2.3.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
python-multipart==0.0.6

# Monitoring
sentry-sdk==1.35.0
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

# Additional Dependencies
radon==6.0.1
slowapi==0.1.9
